a2a/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
google-auth>=2.28.0
httpx>=0.27.0
pydantic>=2.6.0
requests>=2.31.0
sse-starlette>=2.0.0
# Slack authorization layer
pyyaml>=6.0
aiohttp>=3.9
